+++ src/Credfeto.Notification.Bot.Twitch/Services/TwitchChat.cs
@@ -49,19 +49,16 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
 
     private bool _connected;
 
-    public TwitchChat(
-        IOptions<TwitchBotOptions> options,
-        IMessageChannel<TwitchChatMessage> twitchChatMessageChannel,
-        IMediator mediator,
-        ITwitchClient twitchClient,
-        ITwitchStreamStateManager twitchStreamStateManager,
-        ILogger<TwitchChat> logger
-    )
+    public TwitchChat(IOptions<TwitchBotOptions> options,
+                      IMessageChannel<TwitchChatMessage> twitchChatMessageChannel,
+                      IMediator mediator,
+                      ITwitchClient twitchClient,
+                      ITwitchStreamStateManager twitchStreamStateManager,
+                      ILogger<TwitchChat> logger)
     {
         this._twitchChatMessageChannel = twitchChatMessageChannel;
         this._mediator = mediator ?? throw new ArgumentNullException(nameof(mediator));
-        this._twitchStreamStateManager =
-            twitchStreamStateManager ?? throw new ArgumentNullException(nameof(twitchStreamStateManager));
+        this._twitchStreamStateManager = twitchStreamStateManager ?? throw new ArgumentNullException(nameof(twitchStreamStateManager));
         this._logger = logger ?? throw new ArgumentNullException(nameof(logger));
         this._options = (options ?? throw new ArgumentNullException(nameof(options))).Value;
         this._client = twitchClient as TwitchClient ?? throw new ArgumentNullException(nameof(twitchClient));
@@ -83,7 +80,9 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
         this._chatReconnected = this.SubscribeToChatReconnections();
 
         // LOGGING
-        this._chatLogMessage = this._logger.IsEnabled(LogLevel.Debug) ? this.SubscribeToChatLogMessages() : null;
+        this._chatLogMessage = this._logger.IsEnabled(LogLevel.Debug)
+            ? this.SubscribeToChatLogMessages()
+            : null;
 
         // CHAT
         this._chatMessageReceived = this.SubscribeToIncomingChatMessages();
@@ -142,81 +141,58 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
         return Task.CompletedTask;
     }
 
+    [SuppressMessage(category: "Meziantou.Analyzer", checkId: "MA0032: Use CancellationToken", Justification = "Using IDisposable instead")]
     private IDisposable SubscribeToOutgoingChatMessages()
     {
-        return this
-            ._twitchChatMessageChannel.ReadAllAsync(CancellationToken.None)
-            .ToObservable()
-            .Delay(d => Observable.Timer(this.CalculateWithJitter(d)))
-            .Where(this.IsConnectedToChat)
-            .Select(message =>
-                Observable.FromAsync(cancellationToken =>
-                    this.PublishChatMessageAsync(twitchChatMessage: message, cancellationToken: cancellationToken)
-                )
-            )
-            .Concat()
-            .Subscribe();
+        return this._twitchChatMessageChannel.ReadAllAsync(CancellationToken.None)
+                   .ToObservable()
+                   .Delay(d => Observable.Timer(this.CalculateWithJitter(d)))
+                   .Where(this.IsConnectedToChat)
+                   .Select(message => Observable.FromAsync(cancellationToken => this.PublishChatMessageAsync(twitchChatMessage: message, cancellationToken: cancellationToken)))
+                   .Concat()
+                   .Subscribe();
     }
 
+    [SuppressMessage(category: "Meziantou.Analyzer", checkId: "MA0032: Use CancellationToken", Justification = "Using IDisposable instead")]
     private IDisposable SubscribeToIncomingChatMessages()
     {
-        return Observable
-            .FromEventPattern<OnMessageReceivedArgs>(
-                addHandler: h => this._client.OnMessageReceived += h,
-                removeHandler: h => this._client.OnMessageReceived -= h
-            )
-            .Select(messageEvent => messageEvent.EventArgs)
-            .Select(e =>
-                Observable.FromAsync(cancellationToken =>
-                    this.OnMessageReceivedAsync(e: e, cancellationToken: cancellationToken)
-                )
-            )
-            .Concat()
-            .Subscribe();
+        return Observable.FromEventPattern<OnMessageReceivedArgs>(addHandler: h => this._client.OnMessageReceived += h, removeHandler: h => this._client.OnMessageReceived -= h)
+                         .Select(messageEvent => messageEvent.EventArgs)
+                         .Select(e => Observable.FromAsync(cancellationToken => this.OnMessageReceivedAsync(e: e, cancellationToken: cancellationToken)))
+                         .Concat()
+                         .Subscribe();
     }
 
+    [SuppressMessage(category: "Meziantou.Analyzer", checkId: "MA0032: Use CancellationToken", Justification = "Using IDisposable instead")]
     private IDisposable SubscribeToChatLogMessages()
     {
-        return Observable
-            .FromEventPattern<OnLogArgs>(
-                addHandler: h => this._client.OnLog += h,
-                removeHandler: h => this._client.OnLog -= h
-            )
-            .Select(messageEvent => messageEvent.EventArgs)
-            .Subscribe(onNext: this.OnLog);
+        return Observable.FromEventPattern<OnLogArgs>(addHandler: h => this._client.OnLog += h, removeHandler: h => this._client.OnLog -= h)
+                         .Select(messageEvent => messageEvent.EventArgs)
+                         .Subscribe(onNext: this.OnLog);
     }
 
+    [SuppressMessage(category: "Meziantou.Analyzer", checkId: "MA0032: Use CancellationToken", Justification = "Using IDisposable instead")]
     private IDisposable SubscribeToChatReconnections()
     {
-        return Observable
-            .FromEventPattern<OnReconnectedEventArgs>(
-                addHandler: h => this._client.OnReconnected += h,
-                removeHandler: h => this._client.OnReconnected -= h
-            )
-            .Select(messageEvent => messageEvent.EventArgs)
-            .Subscribe(onNext: this.OnReconnected);
+        return Observable.FromEventPattern<OnReconnectedEventArgs>(addHandler: h => this._client.OnReconnected += h, removeHandler: h => this._client.OnReconnected -= h)
+                         .Select(messageEvent => messageEvent.EventArgs)
+                         .Subscribe(onNext: this.OnReconnected);
     }
 
+    [SuppressMessage(category: "Meziantou.Analyzer", checkId: "MA0032: Use CancellationToken", Justification = "Using IDisposable instead")]
     private IDisposable SubscribeToChatDisconnection()
     {
-        return Observable
-            .FromEventPattern<OnDisconnectedEventArgs>(
-                addHandler: h => this._client.OnDisconnected += h,
-                removeHandler: h => this._client.OnDisconnected -= h
-            )
-            .Select(messageEvent => messageEvent.EventArgs)
-            .Subscribe(onNext: this.OnDisconnected);
+        return Observable.FromEventPattern<OnDisconnectedEventArgs>(addHandler: h => this._client.OnDisconnected += h, removeHandler: h => this._client.OnDisconnected -= h)
+                         .Select(messageEvent => messageEvent.EventArgs)
+                         .Subscribe(onNext: this.OnDisconnected);
     }
 
+    [SuppressMessage(category: "Meziantou.Analyzer", checkId: "MA0032: Use CancellationToken", Justification = "Using IDisposable instead")]
     private IDisposable SubscribeToChatConnection()
     {
-        return Observable
-            .FromEventPattern<OnConnectedArgs>(
-                addHandler: h => this._client.OnConnected += h,
-                removeHandler: h => this._client.OnConnected -= h
-            )
-            .Select(messageEvent => messageEvent.EventArgs)
-            .Subscribe(onNext: this.OnConnected);
+        return Observable.FromEventPattern<OnConnectedArgs>(addHandler: h => this._client.OnConnected += h, removeHandler: h => this._client.OnConnected -= h)
+                         .Select(messageEvent => messageEvent.EventArgs)
+                         .Subscribe(onNext: this.OnConnected);
     }
 
     private TimeSpan CalculateWithJitter(TwitchChatMessage twitchChatMessage)
@@ -225,11 +201,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
 
         double delay = Jitter.WithJitter((int)priority, GetMaxSeconds(priority));
 
-        this._logger.DelayingMessage(
-            streamer: twitchChatMessage.Streamer,
-            delay: delay,
-            message: twitchChatMessage.Message
-        );
+        this._logger.DelayingMessage(streamer: twitchChatMessage.Streamer, delay: delay, message: twitchChatMessage.Message);
 
         return TimeSpan.FromSeconds(delay);
 
@@ -240,11 +212,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
                 MessagePriority.ASAP => 2 * (int)MessagePriority.ASAP,
                 MessagePriority.NATURAL => 3 * (int)MessagePriority.NATURAL,
                 MessagePriority.SLOW => 6 * (int)MessagePriority.SLOW,
-                _ => throw new ArgumentOutOfRangeException(
-                    nameof(messagePriority),
-                    actualValue: messagePriority,
-                    message: "Unknown message priority"
-                ),
+                _ => throw new ArgumentOutOfRangeException(nameof(messagePriority), actualValue: messagePriority, message: "Unknown message priority")
             };
         }
     }
@@ -256,9 +224,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
 
     private bool IsConnectedToChat(Streamer streamer)
     {
-        return this._client.JoinedChannels.Any(joinedChannel =>
-            StringComparer.OrdinalIgnoreCase.Equals(x: streamer.Value, y: joinedChannel.Channel)
-        );
+        return this._client.JoinedChannels.Any(joinedChannel => StringComparer.OrdinalIgnoreCase.Equals(x: streamer.Value, y: joinedChannel.Channel));
     }
 
     private async Task PublishChatMessageAsync(TwitchChatMessage twitchChatMessage, CancellationToken cancellationToken)
@@ -267,10 +233,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
 
         try
         {
-            if (
-                this._lastMessage.TryGetValue(key: twitchChatMessage.Streamer, out string? lastMessage)
-                && StringComparer.OrdinalIgnoreCase.Equals(x: lastMessage, y: twitchChatMessage.Message)
-            )
+            if (this._lastMessage.TryGetValue(key: twitchChatMessage.Streamer, out string? lastMessage) && StringComparer.OrdinalIgnoreCase.Equals(x: lastMessage, y: twitchChatMessage.Message))
             {
                 if (!twitchChatMessage.IsCommand)
                 {
@@ -280,11 +243,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
 
             this._lastMessage.TryRemove(key: twitchChatMessage.Streamer, value: out _);
 
-            this._logger.SendingMessage(
-                streamer: twitchChatMessage.Streamer,
-                viewer: this._options.Authentication.Chat,
-                message: twitchChatMessage.Message
-            );
+            this._logger.SendingMessage(streamer: twitchChatMessage.Streamer, viewer: this._options.Authentication.Chat, message: twitchChatMessage.Message);
 
             this._client.SendMessage(channel: twitchChatMessage.Streamer.Value, message: twitchChatMessage.Message);
 
@@ -292,11 +251,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
         }
         catch (Exception exception)
         {
-            this._logger.FailedToSendMessage(
-                streamer: twitchChatMessage.Streamer,
-                message: exception.Message,
-                exception: exception
-            );
+            this._logger.FailedToSendMessage(streamer: twitchChatMessage.Streamer, message: exception.Message, exception: exception);
         }
         finally
         {
@@ -363,21 +318,11 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
         {
             string message = e.ChatMessage.Message;
 
-            await this.HandleChatMessageAsync(
-                streamer: streamer,
-                viewer: viewer,
-                message: message,
-                cancellationToken: cancellationToken
-            );
+            await this.HandleChatMessageAsync(streamer: streamer, viewer: viewer, message: message, cancellationToken: cancellationToken);
         }
     }
 
-    private async Task HandleChatMessageAsync(
-        Streamer streamer,
-        Viewer viewer,
-        string message,
-        CancellationToken cancellationToken
-    )
+    private async Task HandleChatMessageAsync(Streamer streamer, Viewer viewer, string message, CancellationToken cancellationToken)
     {
         try
         {
@@ -386,11 +331,7 @@ public sealed class TwitchChat : ITwitchChat, IDisposable
         }
         catch (Exception exception)
         {
-            this._logger.FailedToHandleChatMessage(
-                streamer: streamer,
-                message: exception.Message,
-                exception: exception
-            );
+            this._logger.FailedToHandleChatMessage(streamer: streamer, message: exception.Message, exception: exception);
         }
     }
 }